# perf(tests): wrap main test functions