--- a/test/main.dart
+++ b/test/main.dart
@@ -1,0 +1,5 @@
+import 'utils/main.dart' as utils_test;
+
+void main() {
+  utils_test.main();
+}