--- a/test/utils/main.dart
+++ b/test/utils/main.dart
@@ -1,0 +1,17 @@
+import 'date_time_extension_test.dart' as date_time_extension_test;
+import 'iterable_extension_test.dart' as iterable_extension_test;
+import 'iterable_string_extension_test.dart' as iterable_string_extension_test;
+import 'num_extension_test.dart' as num_extension_test;
+import 'reg_exp_match_extension_test.dart' as reg_exp_match_extension_test;
+import 'string_extension_test.dart' as string_extension_test;
+import 'time_of_day_extension_test.dart' as time_of_day_extension_test;
+
+void main() {
+  date_time_extension_test.main();
+  iterable_extension_test.main();
+  iterable_string_extension_test.main();
+  num_extension_test.main();
+  reg_exp_match_extension_test.main();
+  string_extension_test.main();
+  time_of_day_extension_test.main();
+}