… test/utils/date_time_extension_test.dart
… test/utils/iterable_extension_test.dart
… test/utils/iterable_string_extension_test.dart
… test/utils/num_extension_test.dart
… test/utils/reg_exp_match_extension_test.dart
… test/utils/string_extension_test.dart
… test/utils/time_of_day_extension_test.dart