This change reduces test execution between files **by ~36%**, as pointed out in https://github.com/flutter/flutter/issues/86722#issuecomment-914688482, although it still keeps practically the same known cold start (https://github.com/flutter/flutter/issues/17130).

| Method       | Tests | Cold start |     Run |        Total |
| ------------ | ----: | ---------: | ------: | -----------: |
| **Per file** |    49 |   15.864 s | 6.508 s | **22.372 s** |
| **Wrapper**  |    49 |   13.714 s | 0.623 s | **14.337 s** |

<img width="500" src="https://quickchart.io/chart?h=180&c={type:%22horizontalBar%22,data:{labels:[%22Per%20file%22,%22Wrapper%22],datasets:[{label:%22Cold%20start%22,data:[15.864,13.714]},{label:%22Run%22,data:[6.508,.623]}]},options:{scales:{xAxes:[{stacked:!0,scaleLabel:{display:!0,labelString:%22seconds%22}}],yAxes:[{stacked:!0}]}}}">

### References

- [GitHub workflow run — Per file](https://github.com/albertms10/cabin_booking/runs/6056589098?check_suite_focus=true#step:9:26)
- [GitHub workflow run — Wrapper](https://github.com/albertms10/cabin_booking/runs/6056654799?check_suite_focus=true#step:9:24)